+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.2.0-dev
 
 - Enhancements
+  - Allows retrieving the shape of a query
   - Uses `:hackney` instead of `:inets`
 
 ## v0.1.0 (2015-06-04)